+++ src/uni_modules/uview-plus/components/u-calendar/calendar.js
@@ -41,6 +41,7 @@ export default {
 		monthNum: 3,
         monthSwitch: false,
         showToday: true,
+        rangeResultMode: 'all',
         weekText: [t("up.week.one"), t("up.week.two"), t("up.week.three"), t("up.week.four"), t("up.week.five"), t("up.week.six"), t("up.week.seven")],
         forbidDays: [],
         forbidDaysToast: t("up.calendar.disabled"),